requirements.txt
# Python dependencies (alternative implementation)
flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
pyyaml==6.0.1
redis==5.0.1
psycopg2-binary==2.9.9
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Linting and formatting
flake8==6.1.0
black==23.12.1
pylint==3.0.3

# Monitoring
prometheus-client==0.19.0
